Update landVerMain.py
Final program, at least for now. It isn't perfect, there are a significant amount of false negatives. Due to formatting differences between how we store landowner records and how Beacon does, some records are tagged as needing updating when they are in fact, correct. To completely resolve this issue would take company wide effort to re-enter landowner records in our database. However, I have yet to find a false positive and so any record tagged as correct can be trusted. That means that the program, as is, is still useful to me. I can safely skip looking up the property records for database entries that are already tagged as correct, saving several hours of work.